requirements.txt
openai
pytrends
gtts
moviepy
pillow
google-auth
google-auth-oauthlib
google-api-python-client
python-dotenv